feat(stats): update stats
Updating bandwidth -> Power and removing Storage. The former is based on
the case for theming consistency (Power, Life, Harmony, Violence) and
makes the game more accessible to a broader audience. The latter is
based on the fact that we no longer need Storage limits since kami
energy will be drained in accordance to its BYTE production