feat(registries): item registry library
Also includes a naming update to the `Capacity` -> `Health` and `Charge` ->
`HealthCurrent` components as well as any components necessary to support
this change. Following this, we should remove `LibRegistry`, `LibModReg`,
`LibModifier` and possibly `LibPrototypes` and update references to
those libraries to this one. Additionally, it's likely we want to
deprecate `LibBattery` to roll up calculations of health and death into
`LibPet` and Food creation into `LibRegistryItem`.

This does not include updates to the ABI for the purpose of making these
sweeping changes a bit easier to read. Will follow up with those and any
fixes to address compilation errors shortly