chore(food): replace LibBattery
Removes all references to LibBattery and implements the necessary
functions in LibPet and LibRegistryItem to support food in the economy
and the feeding of pets.

This does not cap the pet's energy to the total battery life just for
the sake of preserve the sanity of anyone reviewing this code. Those
changes will be included next.

There are some additional changes here like the renaming of functions
and specifically IUintComp -> IComponents in LibPet to standardize the
naming in our codebase. No strong opinions on the latter renaming,
though i do lean towards IComponents for consistency with IWorld. The
only goal here is to be consistent across libraries, and it should be
easy to flip back and forth.