feat(health): drain pet health based on production
The purpose of this update is for better economic balancing. It includes
updates across the pet and production libraries to handle the new
calculations as well as new functions to better delineate
responsibilities among existing library functions. A new state for
ALIVE/DEAD as well as associated functions have been introduced, though
not yet used.