Convert isLiquidatableBy() to External
We are converting this function to external to free up significant space
on our contract size limitations. This should unlock the ability to do
so for other library functions are `view` only.

This includes a "load bearing" test contract now to force a compilation
error during deployment. This is the only way to avoid the critical
foundry error that causes the whole deployment to fail.. don't ask me
why, just trust the process