Trait Rolling Fix and Refinements (#198)
This update moves trait weight interpretation to the SC side for the sake of rarity
legibility and wipes out large swaths of redundant code in the trait generation and
interpretation. This should reduce the gas load on the reveal step by a fair amount
by reducing the number of intermediate storage variable updates. Another world
testnet deployment is also introduced in this PR to support these changes.