New Account Fields (#429)
Support both an auto-incremented `IndexAccount`, `FavoriteFood` and 
`TimeStart` components on `Accounts`. This should then allow us to reference 
accounts by their index for some added early adopter hype and FE opps/ux.

The updated FE call to `AccountRegister` has been populated with a dummy
'doot' value for the food input for now and will be updated during onboarding flow 
refinements.

This will require a redeployment of:
- `IndexAccountComponent`
- `FavoriteFoodComponent`
- `AccountRegisterSystem`

as well as a data migration to remedy `IndexAccount` on existing accounts. need
to discuss the ideal trigger/resolution. this isn't urgent and will be good practice for 
you @chlwys (#406 for reference)