fix fe health drain (#494)
updating the default value of getBonusValue() broke health calcs on the fe whenever bonus is not
set as null coalescers (??) in javascript treat 0 as a non-null value. sets the default value back to
undefined to fix this. includes comment to warn of this and renames index->entityIndex in
kami shape file for clarity as well