Environment mode override option (through url params) (#527)
allows us to override the network the client is connected to based on url params. any
options provided by the `chainConfigs` map in our `constants/chains.ts` file are
considered valid (i.e. `kamigotchi.io/?mode=[ DEV | TEST | PROD ]`)

it's a bit messy atm,, as the code relies on both logic found in `chains.ts` as well as
logic found in `network/config.ts` as the "source of truth". this logical flow needs to 
be consolidated down the line, so that the clientside `Validators` strictly rely on info
accessed through the network object

behavior:
- if valid override provided, use override
- if invalid of no override provided, use environment
- if no environment mode, default to DEV env