Skill Matrix Edges (#539)
there's a rendering issue, where the edges are misaligned prior to
bounding rects being fully resolved through the Nodes and Parent of the
skill tree. unfortunately there's some uncertainty in sequencing of the
data when hot reloading locally. sometimes it will misfire. sometimes
it won't. sometimes it renders perfectly.

now, i probably sound like a madman, but at this point i'm almost
certain. the sequencing of data/render resolution is wholly dependent
on the order in which i saved the files..

i assume this is due to some race condition within the render loop, that's
broken by the place of the files in memory. in local environments, hot
file-reloading allows us to have some control over the sequence. in
prod.. we will not be so lucky. i'm somewhat at a loss for words.
hopefully this just ends up being a non-problem?