update skill deployment sheets and scripts (#683)
this is a first effort towards a much more ironed out deployment from a set of
raw notion databases in more human-friendly format to a nested onchain shape.

not every shape will translate as cleanly as skills do; one-to-many
representations in particular will likely be a headache. but we should make an
effort to follow this pattern for the sake of sanity on both game design and
engineering