skill matrix refactor (#691)
in terms of functional updates, this grants support for direct references to clientside
image assets, which fall back to `ImageURI` on the onchain Entity. this also includes 
a refactor on the skill matrix for.. sanity and breaks out the key logical/styling pieces
of `ItemIcon` into a (skill) `Node` react component to thin down the layered DOM
elements

this thinning removes the mutual exclusivity lines and introduces a clearer visualization
of skill tree tiers. this should prep the skill matrix for further upgrades