more skill modal cleanups. fix-ish interactions (#692)
minimizes the passed data from one layer of the `KamiModal` to the next,
specifically for the Skill DOM tree. this does not yet properly clean up the
Details Component, but it does insure that data on the modal is updated on
demand by modifying `WaitForActionCompletion` to include the pending
action status (since embedded wallets seem to be unreliable at triggering
the final 'complete' state update to actions).

to remedy the update potentially triggering prior to the data being presented
to the FE through our rpc, we hardcode an additional 2s buffer on the wait 
time. this may potentially break the flow for other ui components that rely
on this function, but we can remedy those by baking in a similar buffer.

this is not a long term fix but rather a patch to address similar failures across
the codebase (e.g. with lootbox reveals). what we should do instead is 
unravel our `createActionSystem` code to properly handle the entirety of the
tx event flow without issue

also this fixes the remainder of the broken images on our skill modal