remove account optimization block (#894)
because the `Requests` react component is rendered conditionally, any data we pull 
will only trigger when it's visible. basically, the DOM tree preemptively parses out any 
logic to process based on what needs to be rendered

the only time we actually have this issue is when we process data Prior to the root
rendering of a modal (i.e. directly on the main modal components)