Eternal Gradual Dutch Auctions (#1430)
## Summary

these changes support reroll and gacha ticket auctions directly in the gacha modal. these auctions will run on onyx and musus respectively. this includes a new formulation for the GDA that allows for perpetual activity without resets and overflows

there is some refinement called for here, but in the interest of maintaining collective sanity it seems wise to cut this initial feature release and follow up support for historic pricing and reroll display refactoring in a separate PR 

## Deployments

components
- period

systems
- AuctionBuy
- _AuctionRegistry

data 
- (item) reroll tickets
- (item) onyx
- (auction) gacha tickets gda
- (auction) reroll tickets gda