clean up world api (#1762)
update the world api and its `genCall()` function to include a more explicit `callData` return. `callData` aggregation is now handled one level up

this change should improve readability and make this api more composable for future work


## Possible Concerns
the `gen` endpoint in `world/api/index.ts` returns the `generateCallData` function underneath, which no longer handles the `callData` stacking and only generates the call data string. it seems the only place this endpoint is used in the codebase is `genSingleCall()` in `worldIniter.ts`, which isnt referenced anywhere


going to merge this for now, so i can build on top of it without risk of merge conflicts but tagging you @chlwys, so you can apply any necessary fixes as needed