re-enable auto gacha reveal (#2165)
re-enables auto gacha reveal at a *good enough* level. This logic is not extra robust and will not be able to handle edge cases, but not a prio to fix this rn.

code errs on the side of caution - only sends 1 tx at a time, and will not send if something breaks. Users are very prompted to do this reveals manually anyway so it isn't a big deal if auto-reveal breaks